Revert "add gamecontrol"
This reverts commit 21cdf0790954aede5e6f6c8017016da44909415e.

+++ Tenebra/Assets/Scripts/LevelClass.cs
@@ -8,7 +8,7 @@ public class LevelClass : GenericSkillClass
         nextExpLevel = Formula(currentLevel);
     }
     //Formula dos Leveis
-    public override double Formula(int level)
+    public override long Formula(long level)
     {
         if (level <= 1)
             return 50;
